requirements-colab.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.4.0
PyYAML>=6.0.2
matplotlib>=3.9.0
paddleocr>=2.8.1
easyocr>=1.7.1
